implement logic for deleting authors from DB.
implement MusicDataEditor script for deleting data from DB.
save the MusicDataEditor state to the separate .json file.
configure UI buttons for deleting/updating authors in the list box.

ToDo :
separate logic for deleting and adding data to the DB.
implement commands for UI elements and change the data context for the view.
refactore the code to clearly follow MVVM pattern.