start refactoring the code to follow the MVVM pattern.
create new properties in MainViewModel.cs for binding UI elements.
create the RelayCommand script without realisation for future commands.
create the DBToolsManager to provide MainViewModel access to the model.
create DBManagementKit to provide a generic tool for working with data of the same type (via enums).
refactore the DataProvider to make it a generic reusable class for working with different data.
delete DBDataProviderBase and use a generic DataProvider with DBToolBase base class.
delete all logic from MainWindow.cs and create the MainWindow.xaml on application start.

 ToDo :
 get rid of DBDataEditorBase and MusicDataEditor,
 create generic reusable classes instead of them.
 configure saving/loading logic for DataProvider and add default queries.
 make all the past app functions to work as before refactoring.