add the ability to remove music from the DB, refactore some scripts.
create the DBQueriesContainer to save DBManagementKit queries to the json.
create the DataRemover and configure it to delete the data from DB.
implement the RelayCommand and use it to implement the delete command for the button.
update the DefaultQueries and use the DBQueriesContainer in it to store different queries.
update the DBManagementKit and add a generic method to get different DB tools.
update the ManagementKits script to add a new function for deleting data from the DB.
update the main view and bind new properties to list boxes and buttons.