add functionality to UI buttons, refactore some scripts.
add the ability to remove authors and author's music from the DB.
add the ability to add new music, authors and connect music to authors.
create the DataAdder script to provide the ability to add new content to the DB.
create the DataTableUtility script to provide some new methods for working with DataTable.
create the DataFilters script to provide methods to generate filters for the DataTable.
refactore DataProvider and DataRemover scripts : move the common logic to the DBToolBase.
refactore all methods to use string params instead of integers.
update the DefaultQueries script : add new default queries for adding/deleting data.
update the UI : bind new commands to buttons, swap the position of some buttons.
delete DBDataEditorBase and MusicDataEditor scripts.