add the ability to replace the main info of the selected item, update UI.
create the DataReplacer tool to replace the existing data in the DB.
update the DBManagementKit script to use the created DataReplacer tool.
update the DefaultQueries script : add some default queries for the data replacement.
update the MainViewModel script : add the command for the data replacement.
recreate the main application window : use the grid for better design and responsiveness.